Fix issue with protonating charged ARG
OB was placing a third H on NH1 using a random bond vector.  The
AssignBonds, which is suppose to match known protein residue bonding
patterns, was not working because atom names were not stripped of
whitespace. Also, the original pattern in resdata.txt does not match the
convention of the double bond going to NH1 (it has NH2). Then, even if
AssignBonds sets the right bond order, the
PerceivedBonds fragment matching code undoes it, so refined the bond
typer patterns.